Added tabs block, added transition
