+++ blocks/network-item/network-item.css
@@ -155,9 +155,9 @@ div.network-item div.our-hr a::after,
 div.network-item.image div.text-content-wrapper div.website a::after,
 div.network-item.video div.text-content-wrapper div.website a::after{
     content: url("../../icons/angle-right-blue.svg");
-    position:relative;
+    position: relative;
+    top: 0.1rem;
     left: 0.3rem;
-    top: 0.5rem;
 }
 
 @media screen and (max-width: 62rem) {
